feat: AOP - Logging Security ErrorHandler
Add AOP ErrorHandler to advise on how to handle exceptions.
Update AOP Logging to include logging at the service layer.
Update AOP Security to include advising on Authentication.
Add AppUtil with methods to log stack trace to log.
Fix bug in authentication, needed to add getUserByEmail service call.
Introduce use of ENVIRONMENT variable pulled from .env.
Refactor code mostly cleanup.

+++ src/main/java/dev/tdz/controllers/AppUserController.java
@@ -1,16 +1,18 @@
 package dev.tdz.controllers;
 
-import dev.tdz.aspects.Authorized;
+import dev.tdz.aspects.Authorization;
+import dev.tdz.aspects.ErrorHandler;
 import dev.tdz.aspects.Logging;
 import dev.tdz.entities.AppUser;
-import dev.tdz.exceptions.NotFoundException;
+import dev.tdz.exceptions.NotAuthorizedException;
 import dev.tdz.services.AppUserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Set;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
@@ -22,50 +24,86 @@ public class AppUserController {
     AppUserService appUserService;
 
     @Logging
+    @ErrorHandler
     @GetMapping("/users")
     public Set<AppUser> retrieveAllUsers() {
-        return this.appUserService.getAllAppUsers();
+        Set<AppUser> users = this.appUserService.getAllAppUsers();
+        // Remove password before sending back response.
+        for (AppUser user : users) {
+            user.setPassword(null);
+        }
+        return users;
     }
 
     @Logging
+    @ErrorHandler
     @GetMapping("/users/{id}")
     public AppUser getUserById(@PathVariable int id) {
-        AppUser user;
-        try {
-            user = this.appUserService.getAppUserById(id);
-        } catch (NotFoundException e) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
-        }
+        AppUser user = this.appUserService.getAppUserById(id);
+        // Remove password before sending back response.
+        user.setPassword(null);
         return user;
     }
 
     @Logging
-    @Authorized
+    @ErrorHandler
     @PostMapping("/users")
     public AppUser createUser(@RequestBody AppUser user) {
+        // Override fields of Transient object.
+        user.setId(0);
         this.appUserService.createAppUser(user);
+        user.setPassword(null);
         return user;
     }
 
     @Logging
-    @Authorized
+    @Authorization
+    @ErrorHandler
     @PutMapping("/users/{id}")
     public AppUser updateUser(@PathVariable int id, @RequestBody AppUser user) {
+        // Check existence of entity to be updated.
+        AppUser currUser = this.appUserService.getAppUserById(id);
+        // Get JWT claims.
+        HttpServletRequest request = (
+                (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()
+        ).getRequest();
+        Integer authUserId = (Integer) request.getAttribute("authUserId");
+        // Check if user is updating himself/herself.
+        if (authUserId != id) {
+            throw new NotAuthorizedException("Access denied");
+        }
+        // Override fields of Transient object.
         user.setId(id);
+        // We need to be able to send null as password from the frontend.
+        // This means we do not want to change the pw.
+        // Setting this to the current pw will prevent Spring/Hibernate
+        // from updating this field due to the @DynamicUpdate annotation.
+        if (user.getPassword() == null) {
+            user.setPassword(currUser.getPassword());
+        }
         this.appUserService.updateAppUser(user);
+        // Remove password before sending response back.
+        user.setPassword(null);
         return user;
     }
 
     @Logging
-    @Authorized
+    @Authorization
+    @ErrorHandler
     @DeleteMapping("/users/{id}")
     public Boolean deleteUserById(@PathVariable int id) {
-        Boolean result;
-        try {
-            result = this.appUserService.deleteAppUserById(id);
-        } catch (NotFoundException e) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
+        // Check existence of entity to be deleted.
+        AppUser appUser = this.getUserById(id);
+        // Get JWT claims.
+        HttpServletRequest request = (
+                (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()
+        ).getRequest();
+        Integer authUserId = (Integer) request.getAttribute("authUserId");
+        // Check if user is deleting themselves.
+        if (authUserId != id) {
+            throw new NotAuthorizedException("Access denied");
         }
-        return result;
+        return this.appUserService.deleteAppUserById(id);
     }
+
 }
